Added AgentDynamic abstract base class;
